+++ packages/expo-image/android/src/main/java/expo/modules/image/events/ImageLoadEvent.kt
@@ -5,13 +5,13 @@ import com.bumptech.glide.load.DataSource
 import com.facebook.react.bridge.Arguments
 import com.facebook.react.bridge.ReadableMap
 import com.facebook.react.uimanager.events.Event
-import com.facebook.react.uimanager.events.RCTEventEmitter
+import com.facebook.react.uimanager.events.RCTModernEventEmitter
 import expo.modules.image.enums.ImageCacheType.Companion.fromNativeValue
 
 class ImageLoadEvent(viewId: Int, private val mModel: Any, private val mDataSource: DataSource, private val mBitmapOptions: BitmapFactory.Options) : Event<ImageLoadEvent>(viewId) {
   override fun getEventName() = EVENT_NAME
 
-  override fun dispatch(rctEventEmitter: RCTEventEmitter) {
+  override fun dispatch(rctEventEmitter: RCTModernEventEmitter) {
     val eventData = Arguments.createMap().apply {
       putInt("cacheType", fromNativeValue(mDataSource).enumValue)
       putMap("source", serializeSource(mBitmapOptions, mModel))